requirements.txt
# Requires Python >= 3.12
numpy
pandas
PyYAML
matplotlib
fastparquet
scipy
